+++ ntex-http/CHANGES.md
@@ -2,4 +2,8 @@
 
 ## [0.1.0] - 2022-06-20
 
+* impl TryFrom<&str> for Value
+
+* impl FromIterator for HeaderMap
+
 * Move http types to separate crate